Added a release version
This includes the first usable version of the maze solver which uses a matrix to store nodes and edges which makes the algorithms run at O(n^2) on every new node. But therefore edges are added incredibly fast so the BFS runs fast but the Initialization takes a lot of time. I need to implement a graph that uses adjacent lists to balance this more out